Moving plugin to StatesDumper
