--- Persistence/Persistence.csproj
@@ -1,21 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <Version>2.9.1</Version>
-    <TargetFrameworks>netstandard2.0;net47</TargetFrameworks>
-    <RootNamespace>Neo.Plugins</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Update="Persistence\config.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </None>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Neo" Version="2.9.1" />
-
-  </ItemGroup>
-
-</Project>
