--- Persistence/Persistence.sln
@@ -1,17 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2017
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence", "Persistence.csproj", "{AFBDCF62-12B5-4EEA-A309-ECF772AD1985}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{AFBDCF62-12B5-4EEA-A309-ECF772AD1985}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AFBDCF62-12B5-4EEA-A309-ECF772AD1985}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AFBDCF62-12B5-4EEA-A309-ECF772AD1985}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AFBDCF62-12B5-4EEA-A309-ECF772AD1985}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
