+++ StatesDumper/Settings.cs
@@ -9,11 +9,11 @@ namespace Neo.Plugins
 {
     internal class Settings
     {
-	    //Amount of storages states (heights) to be dump in a given file
+	///Amount of storages states (heights) to be dump in a given file
         public uint BlockCacheSize { get; }
-	    //Height to begin storage dump
+	///Height to begin storage dump
         public uint HeightToBegin { get; }
-	    //Height to begin real-time syncing and dumping on single files
+	///Height to begin real-time syncing and dumping on single files
         public uint HeightToStartRealTimeSyncing { get; }
         public uint PersistAction { get; }
         public string BlockStorageCache;
@@ -29,6 +29,7 @@ namespace Neo.Plugins
 
         public Settings(IConfigurationSection section)
         {
+	    /// Geting settings for storage changes state dumper
             this.BlockCacheSize = GetValueOrDefault(section.GetSection("BlockCacheSize"), 1000u, p => uint.Parse(p));
             this.HeightToBegin = GetValueOrDefault(section.GetSection("HeightToBegin"), 0u, p => uint.Parse(p));
             this.HeightToStartRealTimeSyncing = GetValueOrDefault(section.GetSection("HeightToStartRealTimeSyncing"), 2883000u, p => uint.Parse(p));
