+++ StatesDumper/StatesDumper.sln
@@ -1,0 +1,17 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2017
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StatesDumper", "StatesDumper.csproj", "{018AC353-72BA-4DC7-B551-1865C4531373}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{018AC353-72BA-4DC7-B551-1865C4531373}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{018AC353-72BA-4DC7-B551-1865C4531373}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{018AC353-72BA-4DC7-B551-1865C4531373}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{018AC353-72BA-4DC7-B551-1865C4531373}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
